Move search_engine to TBIR_app
