+++ TBIR_app/views.py
@@ -1,3 +1,11 @@
 from django.shortcuts import render
 
-# Create your views here.
+
+def welcome(request):
+    return render(request, 'index.html')
+
+
+def result(request):
+    index = request.GET['index']
+    print("Fix here for retrieve the photo")
+    return render(request, 'result.html')
