--- search_engine/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class SearchEngineConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'search_engine'
